initListIterator

+++ lab1/Регистр сдвига/Graph.h
@@ -13,8 +13,11 @@ struct Graph {
 void enablePrintSearchLog();
 int initGraph(struct Graph *graph, uint64_t num_of_nodes);
 void freeGraph(struct Graph *graph);
-uint8_t edgeExists(struct Graph *graph, uint32_t node1, uint32_t node2);
-void setOrDeleteEdge(struct Graph *graph, uint32_t node1, uint32_t node2, uint8_t set);
+uint8_t edgeExists(struct Graph *graph, uint64_t node1, uint64_t node2);
+void setOrDeleteEdge(
+    struct Graph *graph,
+    uint64_t node1, uint64_t node2, uint8_t set
+);
 int getComponents(struct Graph *graph, List *components);
 void makeUnoriented(struct Graph *graph);
 int reverseGraph(struct Graph *original, struct Graph *reversed);
